[project @ 2003-02-25 19:01:12 by panne] Say goodbye to completely automatically generated enumerants and their (un-)marshalers...
The reason for this probably surprising move is that it is impossible
to generate good documentation from .spec-generated Haskell code
without heavy modification of the .spec files, which would be
something very undesirable. Furthermore, the values of the enumerants
are nailed down across platforms by a heavily standardized process
through the extension registry, so we don't suffer any
incompatibilities.

Note that we still generate enumerants automatically, but they are
inserted where they are needed via cut-n-paste. Sounds very low-tech,
but we have to add extensive documentation by hand, anyway, so nothing
is really lost.

It was a long way to this seemingly simple and obvious way of handling
enumerants in HOpenGL, but I think I've finally figured out the
"right" way, once and forever. (famous last words... :-}