[project @ 2003-06-29 15:51:06 by panne] After some serious fights with the OpenGL spec it was concluded that defending against every possible abuse of the API with the help of Haskell's type system is a futile attempt: There are cases where the allowed values of one parameter depend on the value of another, e.g. in the area of pixel formats and pixel types. Consequently, types which are used for similar purposes are merged, e.g. the buffer modes for reading and writing, although the former is only a subset of the latter.
Although this allows for a few semantic errors which were previously
impossible, the new design is more consistent and quite a bit simpler.

Furthermore, the strucure of the source code was cleaned up by moving
every data types which is used more than once (including its
(un-)marshalers) into a separate module of the same name.