Splits the PName into parts for typesafety
The types are devided into categories based on the size of the return
type (1,2,3,4 or N), the type of values (integer or floating) and
whether it can be used unindexed, indexed or both.
Possibly boolean queries should be turned of. Several old PName values
were not used and are left commented out in the current PName.